Split traffic_accidents.rs (590 lines) into subdirectory modules
Split crates/simulation/src/traffic_accidents.rs into focused modules
under traffic_accidents/ for better maintainability:

- types.rs (44 lines): TrafficAccident, AccidentTracker structs
- calculations.rs (34 lines): road_neighbor_directions, weather_accident_multiplier
- systems.rs (264 lines): spawn_accidents, process_accidents, TrafficAccidentsPlugin
- tests.rs (256 lines): all unit tests
- mod.rs (7 lines): thin facade with re-exports

All public items re-exported from mod.rs — no external import changes needed.
